refactor(pos-mcp-server): move RagScope to internal.domain for cross-layer sharing
