+++ pos-mcp-server/src/main/java/com/positivity/mcp/internal/orchestration/rag/ScopedContentRetrieverFactory.java
@@ -2,6 +2,7 @@ package com.positivity.mcp.internal.orchestration.rag;
 
 import static dev.langchain4j.store.embedding.filter.MetadataFilterBuilder.metadataKey;
 
+import com.positivity.mcp.internal.domain.RagScope;
 import dev.langchain4j.model.embedding.EmbeddingModel;
 import dev.langchain4j.rag.content.retriever.ContentRetriever;
 import dev.langchain4j.rag.content.retriever.EmbeddingStoreContentRetriever;
